Refactor onboarding and settings flow
