+++ ios/Cove/Flows/SettingsFlow/MainSettingsScreen.swift
@@ -235,216 +235,6 @@ struct MainSettingsScreen: View {
         }
     }
 
-    @ViewBuilder
-    var CloudBackupSection: some View {
-        if !auth.isInDecoyMode() {
-            let manager = CloudBackupManager.shared
-
-            Section(header: Text("Cloud Backup")) {
-                switch manager.lifecycle {
-                case .disabled:
-                    SettingsRow(title: "Enable Cloud Backup", symbol: "icloud.and.arrow.up") {
-                        sheetState = .init(.cloudBackupOnboarding)
-                    }
-                case .enabling:
-                    cloudBackupEnablingRow
-                case .restoring:
-                    cloudBackupRestoringRow
-                case let .failed(failure):
-                    cloudBackupErrorContent(message: failure.message)
-                case .configured:
-                    cloudBackupEnabledRow(manager: manager)
-                }
-            }
-        }
-    }
-
-    private var cloudBackupEnablingRow: some View {
-        HStack {
-            SettingsIcon(symbol: "icloud.and.arrow.up")
-            Text("Setting up cloud backup...")
-                .font(.subheadline)
-                .padding(8)
-            Spacer()
-            ProgressView()
-        }
-    }
-
-    private func cloudBackupEnabledRow(manager: CloudBackupManager) -> some View {
-        HStack {
-            cloudBackupEnabledStatus(manager: manager)
-            Spacer()
-            settingsChevron
-        }
-        .contentShape(Rectangle())
-        .onTapGesture {
-            app.pushRoute(Route.settings(.cloudBackup))
-        }
-    }
-
-    @ViewBuilder
-    private func cloudBackupEnabledStatus(manager: CloudBackupManager) -> some View {
-        let status = manager.settingsRowStatus
-
-        switch status {
-        case .disabled, .disabling, .settingUp, .restoring:
-            cloudBackupStatusContent(
-                symbol: "icloud",
-                title: "Cloud Backup",
-                color: Color.secondary
-            )
-
-        case .passkeyMissing:
-            cloudBackupStatusContent(
-                symbol: "exclamationmark.icloud.fill",
-                title: "Cloud Backup Passkey Missing",
-                message: "Backups can't be restored until you add a new passkey",
-                color: Color.statusWarning
-            )
-
-        case .passkeyProviderUnsupported:
-            cloudBackupStatusContent(
-                symbol: "exclamationmark.icloud.fill",
-                title: "Cloud Backup Passkey Unsupported",
-                message: "Open to choose a supported passkey provider",
-                color: Color.statusWarning
-            )
-
-        case .unverified:
-            cloudBackupStatusContent(
-                symbol: "exclamationmark.icloud",
-                title: "Cloud Backup Unverified",
-                color: Color.statusWarning
-            )
-
-        case .confirming:
-            cloudBackupStatusContent(
-                symbol: "arrow.clockwise.icloud",
-                title: "Cloud Backup Confirming",
-                color: Color.statusInfo
-            )
-
-        case .active:
-            cloudBackupStatusContent(
-                symbol: "checkmark.icloud",
-                title: "Cloud Backup Enabled",
-                color: Color.statusSuccess
-            )
-
-        case .verificationRecommended:
-            cloudBackupStatusContent(
-                symbol: "exclamationmark.icloud",
-                title: "Cloud Backup Enabled",
-                message: "Verification recommended",
-                color: Color.statusWarning
-            )
-
-        case .checkingSync:
-            cloudBackupStatusContent(
-                symbol: "icloud",
-                title: "Checking Cloud Backup",
-                message: "Checking iCloud sync status",
-                color: Color.secondary
-            )
-
-        case .syncing:
-            cloudBackupStatusContent(
-                symbol: "arrow.clockwise.icloud",
-                title: "Cloud Backup Syncing",
-                message: "Uploading latest changes",
-                color: Color.statusInfo
-            )
-
-        case .noFiles:
-            cloudBackupStatusContent(
-                symbol: "icloud.slash",
-                title: "Cloud Backup Needs Attention",
-                message: "No iCloud backup files found",
-                color: Color.statusWarning
-            )
-
-        case .driveUnavailable:
-            cloudBackupStatusContent(
-                symbol: "exclamationmark.icloud",
-                title: "iCloud Drive Unavailable",
-                message: "Open to review Cloud Backup",
-                color: Color.statusWarning
-            )
-
-        case let .authorizationRequired(message):
-            cloudBackupStatusContent(
-                symbol: "exclamationmark.icloud",
-                title: "iCloud Access Needed",
-                message: message,
-                color: Color.statusWarning
-            )
-
-        case let .error(message):
-            cloudBackupStatusContent(
-                symbol: "exclamationmark.icloud",
-                title: "Cloud Backup Error",
-                message: message,
-                color: Color.statusError
-            )
-        }
-    }
-
-    @ViewBuilder
-    private func cloudBackupStatusContent(
-        symbol: String,
-        title: String,
-        message: String? = nil,
-        color: Color
-    ) -> some View {
-        Image(systemName: symbol)
-            .foregroundStyle(color)
-
-        VStack(alignment: .leading, spacing: 2) {
-            Text(title)
-
-            if let message {
-                Text(message)
-                    .font(.caption2)
-                    .foregroundStyle(color)
-                    .lineLimit(1)
-            }
-        }
-    }
-
-    private var cloudBackupRestoringRow: some View {
-        HStack {
-            ProgressView()
-                .padding(.trailing, 8)
-            Text("Restoring from cloud backup...")
-        }
-    }
-
-    private func cloudBackupErrorContent(message: String) -> some View {
-        Group {
-            VStack(alignment: .leading, spacing: 4) {
-                HStack {
-                    Image(systemName: "exclamationmark.icloud")
-                        .foregroundStyle(Color.statusError)
-                    Text("Cloud Backup Error")
-                }
-                Text(message)
-                    .font(.caption)
-                    .foregroundStyle(.secondary)
-            }
-
-            SettingsRow(title: "Review", symbol: "arrow.right") {
-                app.pushRoute(Route.settings(.cloudBackup))
-            }
-        }
-    }
-
-    private var settingsChevron: some View {
-        Image(systemName: "chevron.right")
-            .foregroundColor(Color(UIColor.tertiaryLabel))
-            .font(.footnote)
-            .fontWeight(.semibold)
-    }
-
     var betaToggle: Binding<Bool> {
         Binding(
             get: { isBetaEnabled },
@@ -504,7 +294,15 @@ struct MainSettingsScreen: View {
                 sheetState = .init(.changePin)
             }
             BackupSection
-            CloudBackupSection
+            MainSettingsCloudBackupSection(
+                isVisible: !auth.isInDecoyMode(),
+                onEnable: {
+                    sheetState = .init(.cloudBackupOnboarding)
+                },
+                onOpenDetail: {
+                    app.pushRoute(Route.settings(.cloudBackup))
+                }
+            )
             BetaToggleSection
 
             Section {
@@ -987,199 +785,6 @@ struct MainSettingsScreen: View {
         do { try auth.rust.setDecoyPin(pin: pin) } catch {
             let error = error as! AuthManagerError
             alertState = .init(.extraSetPinError(error.description))
-        }
-    }
-}
-
-private struct SettingsCloudBackupEnableSheet: View {
-    @State private var manager = CloudBackupManager.shared
-    @State private var ignoreNextPromptDismiss = false
-
-    let onComplete: () -> Void
-    let onDismiss: () -> Void
-
-    private var message: String? {
-        if manager.isUnsupportedPasskeyProvider {
-            "This passkey provider did not confirm PRF support for Cloud Backup. Try Apple Passwords (iCloud Keychain) or another supported provider such as 1Password"
-        } else {
-            manager.lifecycleFailureMessage
-        }
-    }
-
-    private var isBusy: Bool {
-        if case .awaitingSavedPasskeyConfirmation(.manual) = manager.enableFlow {
-            return false
-        }
-
-        if isAwaitingEnablePrompt(manager.rootPrompt) {
-            return false
-        }
-
-        return manager.isLifecycleEnabling
-    }
-
-    private var showingPasskeyChoice: Binding<Bool> {
-        Binding(
-            get: { isEnablePasskeyChoice(manager.rootPrompt) },
-            set: { isPresented in
-                guard !isPresented else { return }
-                handlePromptDismiss()
-            }
-        )
-    }
-
-    private var showingExistingBackupPrompt: Binding<Bool> {
-        Binding(
-            get: {
-                if case .existingBackupFound = manager.rootPrompt { return true }
-                return false
-            },
-            set: { isPresented in
-                guard !isPresented else { return }
-                handlePromptDismiss()
-            }
-        )
-    }
-
-    private var passkeyChoiceHint: CloudBackupPasskeyHint? {
-        guard case let .passkeyChoice(.enable(_, passkeyHint)) = manager.rootPrompt else {
-            return nil
-        }
-
-        return passkeyHint
-    }
-
-    private var existingBackupPasskeyHint: CloudBackupPasskeyHint? {
-        guard case let .existingBackupFound(_, passkeyHint) = manager.rootPrompt else {
-            return nil
-        }
-
-        return passkeyHint
-    }
-
-    private func isEnablePasskeyChoice(_ rootPrompt: CloudBackupRootPrompt) -> Bool {
-        guard case let .passkeyChoice(intent) = rootPrompt else { return false }
-        if case .enable = intent { return true }
-        return false
-    }
-
-    private func isAwaitingEnablePrompt(_ rootPrompt: CloudBackupRootPrompt) -> Bool {
-        if case .existingBackupFound = rootPrompt { return true }
-        return isEnablePasskeyChoice(rootPrompt)
-    }
-
-    private func beginEnableChoice() {
-        guard !isBusy, !isAwaitingEnablePrompt(manager.rootPrompt) else { return }
-        manager.dispatch(action: .promptEnablePasskeyChoice(.init(
-            savedPasskeyConfirmation: .manual,
-            verificationSource: .settings
-        )))
-    }
-
-    private func dispatchPromptAction(_ action: CloudBackupManagerAction) {
-        ignoreNextPromptDismiss = true
-        manager.dispatch(action: action)
-    }
-
-    private func handlePromptDismiss() {
-        if ignoreNextPromptDismiss {
-            ignoreNextPromptDismiss = false
-            return
-        }
-
-        switch manager.rootPrompt {
-        case .existingBackupFound:
-            manager.dispatch(action: .discardPendingEnableCloudBackup)
-        case .passkeyChoice(.enable):
-            manager.dispatch(action: .dismissPasskeyChoicePrompt)
-        case .none, .missingPasskeyReminder, .passkeyChoice(.repairPasskey), .verification:
-            break
-        }
-    }
-
-    private func completeIfReady(_ completion: TaggedItem<CloudBackupEnableContext>?) {
-        guard completion?.item.verificationSource == .settings else { return }
-
-        onComplete()
-    }
-
-    private func existingPasskeyButtonTitle(for hint: CloudBackupPasskeyHint?) -> String {
-        guard let hint else { return "Use Existing Passkey" }
-        return "Use Existing Passkey (\(hint.nameSuffix))"
-    }
-
-    private var existingBackupMessage: String {
-        guard let existingBackupPasskeyHint else {
-            return "Creating a new Cloud Backup will not include wallets from your previous backup. If you still have access to the passkey for that backup, use the existing passkey instead."
-        }
-
-        return "Creating a new Cloud Backup will not include wallets from your previous backup. If you still have access to the passkey named Cove Cloud Backup (\(existingBackupPasskeyHint.nameSuffix)), use that passkey instead."
-    }
-
-    var body: some View {
-        ZStack {
-            if case .awaitingSavedPasskeyConfirmation(.manual) = manager.enableFlow {
-                CloudBackupEnableConfirmationView(
-                    onContinue: {
-                        manager.dispatch(action: .confirmSavedPasskey)
-                    },
-                    onCancel: {
-                        manager.dispatch(action: .discardPendingEnableCloudBackup)
-                        onDismiss()
-                    }
-                )
-            } else {
-                CloudBackupEnableOnboardingView(
-                    onEnable: beginEnableChoice,
-                    onCancel: onDismiss,
-                    message: message,
-                    isBusy: isBusy
-                )
-            }
-
-            if isBusy {
-                CloudBackupEnableBusyOverlay(enableFlow: manager.enableFlow)
-            }
-        }
-        .onChange(of: manager.enableCompletion) { _, completion in
-            completeIfReady(completion)
-        }
-        .onChange(of: manager.rootPrompt, initial: true) { _, rootPrompt in
-            if !isAwaitingEnablePrompt(rootPrompt) {
-                ignoreNextPromptDismiss = false
-            }
-        }
-        .alert(
-            "Passkey Options",
-            isPresented: showingPasskeyChoice
-        ) {
-            Button(existingPasskeyButtonTitle(for: passkeyChoiceHint)) {
-                dispatchPromptAction(.acceptEnablePrompt(.useExisting))
-            }
-            Button("Create New Passkey") {
-                dispatchPromptAction(.acceptEnablePrompt(.createNew))
-            }
-            Button("Cancel", role: .cancel) {
-                dispatchPromptAction(.dismissPasskeyChoicePrompt)
-            }
-        } message: {
-            Text("Would you like to use an existing passkey or create a new one?")
-        }
-        .alert(
-            "Existing Cloud Backup Found",
-            isPresented: showingExistingBackupPrompt
-        ) {
-            Button("Create New Backup", role: .destructive) {
-                dispatchPromptAction(.acceptEnablePrompt(.createNew))
-            }
-            Button("Try Existing Passkey") {
-                dispatchPromptAction(.acceptEnablePrompt(.useExisting))
-            }
-            Button("Cancel", role: .cancel) {
-                dispatchPromptAction(.discardPendingEnableCloudBackup)
-            }
-        } message: {
-            Text(existingBackupMessage)
         }
     }
 }
